prototypes/noelle/requirements.txt
discord.py>=2.4.0
discord-ext-voice-recv @ git+https://github.com/imayhaveborkedit/discord-ext-voice-recv.git
deepgram-sdk>=4.0.0
python-dotenv>=1.0.1
PyNaCl>=1.5.0
